Refactor stats into a module
